+++ packages/clarity-visualize/src/clarity.ts
@@ -1,3 +1,3 @@
-import { Visualizer } from "./visualizerx";
+import { Visualizer } from "./visualizer";
 
 export const { state, dom, get, html, time, clickmap, clearmap, scrollmap, merge, setup, render } = new Visualizer();
