+++ packages/clarity-visualize/src/index.ts
@@ -1,2 +1,2 @@
 export * as visualize from "./clarity";
-export { Visualizer } from "./visualizerx";
+export { Visualizer } from "./visualizer";
